+++ src/libs/actions/BankAccounts.js
@@ -3,7 +3,6 @@ import CONST from '../../CONST';
 import * as API from '../API';
 import ONYXKEYS from '../../ONYXKEYS';
 import * as Localize from '../Localize';
-import DateUtils from '../DateUtils';
 
 export {
     setupWithdrawalAccount,
@@ -41,49 +40,6 @@ function clearPersonalBankAccount() {
 function clearPlaid() {
     Onyx.set(ONYXKEYS.PLAID_DATA, {});
     Onyx.set(ONYXKEYS.PLAID_LINK_TOKEN, '');
-}
-
-/**
- * Helper method to build the Onyx data required during setup of a Verified Business Bank Account
- *
- * @returns {Object}
- */
-// We'll remove the below once this function is used by the VBBA commands that are yet to be implemented
-function getVBBADataForOnyx() {
-    return {
-        optimisticData: [
-            {
-                onyxMethod: CONST.ONYX.METHOD.MERGE,
-                key: ONYXKEYS.REIMBURSEMENT_ACCOUNT,
-                value: {
-                    isLoading: true,
-                    errors: null,
-                },
-            },
-        ],
-        successData: [
-            {
-                onyxMethod: CONST.ONYX.METHOD.MERGE,
-                key: ONYXKEYS.REIMBURSEMENT_ACCOUNT,
-                value: {
-                    isLoading: false,
-                    errors: null,
-                },
-            },
-        ],
-        failureData: [
-            {
-                onyxMethod: CONST.ONYX.METHOD.MERGE,
-                key: ONYXKEYS.REIMBURSEMENT_ACCOUNT,
-                value: {
-                    isLoading: false,
-                    errors: {
-                        [DateUtils.getMicroseconds()]: Localize.translateLocal('paymentsPage.addBankAccountFailure'),
-                    },
-                },
-            },
-        ],
-    };
 }
 
 /**
@@ -164,5 +120,4 @@ export {
     deletePaymentBankAccount,
     clearPersonalBankAccount,
     clearPlaid,
-    getVBBADataForOnyx,
 };
